Convert various sizes and lengths to size_t; add error handling for E2BIG on BFQ messages; add ASSUMELARGEMEM for pre-allocating extcombo strbuffer to max size
git-svn-id: svn://svn.code.sf.net/p/xymon/code@7750 44351d6e-118b-4698-b696-ce33095ecaa4

+++ branches/4.x-master/include/libxymon.h
@@ -22,7 +22,7 @@ typedef struct htnames_t {
 
 typedef struct strbuffer_t {
 	char *s;
-	int used, sz;
+	size_t used, sz;
 } strbuffer_t;
 
 #define STRBUF(buf) (buf->s)
